+++ cpp_basics/cpp_eindopdracht/oscillator.h
@@ -6,28 +6,28 @@ class Oscillator
 {
 public:
   //Constructor and destructor
-  Oscillator(float frequency, float samplerate);
+  Oscillator(double frequency, double samplerate);
   ~Oscillator();
 
   //return the current sample
-  float getSample();
+  double getSample();
   // go to next sample
   void tick();
   virtual void calculate() = 0;
 
   //getters and setters
-  void setFrequency(float frequency);
-  float getFrequency();
+  void setFrequency(double frequency);
+  double getFrequency();
 
   //NOTE - do we need a setter for phase? for now -> not using one
 
 protected:
-  float amplitude;
-  float frequency;
-  float phase;
+  double amplitude;
+  double frequency;
+  double phase;
   // contains the current sample
-  float sample;
-  float samplerate;
+  double sample;
+  double samplerate;
 };
 
 #endif
