+++ cpp_basics/cpp_eindopdracht/saw.h
@@ -3,7 +3,7 @@
 class Saw : public Oscillator{
 public:
   // constructor and destructor
-  Saw(float frequency, float samplerate);
+  Saw(double frequency, double samplerate);
   ~Saw();
 
   // methods
