+++ cpp_basics/cpp_eindopdracht/sine.h
@@ -3,9 +3,9 @@
 class Sine : public Oscillator{
 public:
   // constructor and destructor
-  Sine(float frequency, float samplerate);
+  Sine(double frequency, double samplerate);
   ~Sine();
 
   // methods
-  void calculate() override; 
+  void calculate() override;
 };
